DoublyLinkedCircularList.insert() fixed and .__repr__() implemented
Further testing will be needed, but so far it seems that I got .insert() right. The __repr__() function seems to be fine as well. Theoretically, given the purpose of this program, probably I wouldn't need a .remove() method, but I think I'll implement it anyway – it's better to have it and not need it than the other way around.