More documentation and code optimization
The skeleton for a .remove() method is complete, along with minimal documentation.

ATTENTION: Visual Studio is mixing line endings, so Python at this moment throws an Exception when reading some of the new documentation, due to "unknown character", or something of the kind. I'll fix that later. Got classes to attend!