DoublyLinkedPlayerList moved to uno.py.
Since the DoublyLinkedPlayerList was dependent on the UnoPlayer class, they could not remain separated. As such, the DoublyLinkedPlayerList was moved back into uno.py.

As a curious outcome, this file was rendered usess! Not quite useless, though, since the soul of the DoublyLinkedCircularList was adapted and became the DoublyLinkedPlayerList, which – by the way – probably will be renamed later on to DoublyLinkedCircularPlayerList, since it shows in the name the circular nature of the list, which is crucial in its application.